# Per channel quantization performance improvement