--- a/aten/src/ATen/native/quantized/cpu/kernels/QuantizedOpKernels.cpp
+++ b/aten/src/ATen/native/quantized/cpu/kernels/QuantizedOpKernels.cpp
@@ -1080,7 +1080,7 @@ void q_batch_norm_kernel(
 
 }
 
-void fake_quantize_slice_kernel(
+void fake_quantize_tensor_kernel(
     Tensor& output,
     const Tensor& input,
     float sc,
@@ -1101,7 +1101,7 @@ void fake_quantize_slice_kernel(
   });
 }
 
-void fake_quantize_grad_slice_kernel(
+void fake_quantize_grad_tensor_kernel(
     Tensor& input_grad,
     const Tensor& input,
     const Tensor& output_grad,
@@ -1115,6 +1115,30 @@ void fake_quantize_grad_slice_kernel(
     int64_t xq = static_cast<int64_t>(std::nearbyint(x * inv_scale + z_point));
     return dy * (xq >= quant_min && xq <= quant_max);
   });
+}
+
+void fake_quant_by_channel_cuda(TensorIterator &iter, int64_t quant_min, int64_t quant_max) {
+  cpu_kernel(iter,
+    [=](float self, float scale, int64_t zero_point) -> float {
+      float inv_scale = 1.0f / scale;
+      return (std::fmin(
+                std::fmax(
+                    static_cast<int64_t>(
+                        std::nearbyint(self * inv_scale + zero_point)),
+                    quant_min),
+                quant_max) -
+            zero_point) *
+        scale;
+    });
+}
+
+void fake_quant_grad_by_channel_cuda(TensorIterator &iter, int64_t quant_min, int64_t quant_max) {
+  cpu_kernel(iter,
+    [=](float x, float dy, float scale, int64_t zero_point) -> float {
+      float inv_scale = 1.0f / scale;
+      int64_t xq = static_cast<int64_t>(std::nearbyint(x * inv_scale + zero_point));
+      return dy * (xq >= quant_min && xq <= quant_max);
+    });
 }
 
 } // namespace
@@ -1139,8 +1163,10 @@ REGISTER_DISPATCH(qcat_nhwc_stub, &qcat_nhwc_kernel<false>);
 REGISTER_DISPATCH(qcat_relu_nhwc_stub, &qcat_nhwc_kernel<true>);
 REGISTER_DISPATCH(qtopk_stub, &qtopk_kernel);
 REGISTER_DISPATCH(qbatch_norm_stub, &q_batch_norm_kernel<false>);
-REGISTER_DISPATCH(fake_quant_slice_stub, &fake_quantize_slice_kernel);
-REGISTER_DISPATCH(fake_quant_grad_slice_stub, &fake_quantize_grad_slice_kernel);
+REGISTER_DISPATCH(fake_quant_tensor_stub, &fake_quantize_tensor_kernel);
+REGISTER_DISPATCH(fake_quant_grad_tensor_stub, &fake_quantize_grad_tensor_kernel);
+REGISTER_DISPATCH(fake_quant_by_channel_stub, &fake_quant_by_channel_cpu);
+REGISTER_DISPATCH(fake_quant_grad_by_channel_stub, &fake_quant_grad_by_channel_cpu);
 
 } // namespace native
 } // namespace at