--- a/aten/src/ATen/native/quantized/cuda/fake_quantize_core.cu
+++ b/aten/src/ATen/native/quantized/cuda/fake_quantize_core.cu
@@ -4,7 +4,7 @@
 #include <ATen/native/quantized/fake_quant_affine.h>
 #include <cmath>
 
-/* Fake quantize a tensor, common block for per-channel & per-tensor fake quant
+/* Fake quantize a tensor
 Args:
   output: output tensor.
   input : input tensor.
@@ -17,7 +17,7 @@ Returns:
 */
 namespace at {
 namespace native {
-void fake_quantize_slice_kernel_cuda(
+void fake_quantize_tensor_kernel_cuda(
     Tensor& output,
     const Tensor& input,
     float scale,
@@ -38,7 +38,7 @@ void fake_quantize_slice_kernel_cuda(
       });
 }
 
-void fake_quantize_grad_slice_kernel_cuda(
+void fake_quantize_grad_tensor_kernel_cuda(
     Tensor& input_grad,
     const Tensor& input,
     const Tensor& output_grad,
@@ -57,8 +57,37 @@ void fake_quantize_grad_slice_kernel_cuda(
       });
 }
 
-REGISTER_DISPATCH(fake_quant_slice_stub, &fake_quantize_slice_kernel_cuda);
-REGISTER_DISPATCH(fake_quant_grad_slice_stub, &fake_quantize_grad_slice_kernel_cuda);
+REGISTER_DISPATCH(fake_quant_tensor_stub, &fake_quantize_tensor_kernel_cuda);
+REGISTER_DISPATCH(fake_quant_grad_tensor_stub, &fake_quantize_grad_tensor_kernel_cuda);
+
+// Fake quantize by channel
+
+void fake_quant_by_channel_cuda(TensorIterator &iter, int64_t quant_min, int64_t quant_max) {
+  gpu_kernel(iter,
+    [=] GPU_LAMBDA (float input_val, float scale, int64_t zero_point) -> float {
+      float inv_scale = 1.0f / scale;
+      return (fminf(
+                quant_max,
+                fmaxf(
+                    quant_min,
+                    static_cast<int64_t>(std::nearbyint(
+                        input_val * inv_scale + zero_point)))) -
+            zero_point) *
+          scale;
+    });
+}
+
+void fake_quant_grad_by_channel_cuda(TensorIterator &iter, int64_t quant_min, int64_t quant_max) {
+  gpu_kernel(iter,
+    [=] GPU_LAMBDA (float x, float dy, float scale, int64_t zero_point) -> float {
+      float inv_scale = 1.0f / scale;
+      int64_t Xq = std::nearbyint(x * inv_scale + zero_point);
+      return (Xq >= quant_min && Xq <= quant_max) * dy;
+    });
+}
+
+REGISTER_DISPATCH(fake_quant_by_channel_stub, &fake_quant_by_channel_cuda);
+REGISTER_DISPATCH(fake_quant_grad_by_channel_stub, &fake_quant_grad_by_channel_cuda);
 
 } // namespace native
 } // namespace at