--- a/aten/src/ATen/native/quantized/fake_quant_affine.h
+++ b/aten/src/ATen/native/quantized/fake_quant_affine.h
@@ -29,13 +29,13 @@ using fake_quant_grad_tensor_fn = void (*)(
 DECLARE_DISPATCH(fake_quant_tensor_fn, fake_quant_tensor_stub);
 DECLARE_DISPATCH(fake_quant_grad_tensor_fn, fake_quant_grad_tensor_stub);
 
-using fake_quant_by_channel_fn = void (*)(
+using fake_quant_per_channel_fn = void (*)(
     TensorIterator &iter,
     int64_t quant_min,
     int64_t quant_max);
 
-DECLARE_DISPATCH(fake_quant_by_channel_fn, fake_quant_by_channel_stub);
-DECLARE_DISPATCH(fake_quant_by_channel_fn, fake_quant_grad_by_channel_stub);
+DECLARE_DISPATCH(fake_quant_per_channel_fn, fake_quant_per_channel_stub);
+DECLARE_DISPATCH(fake_quant_per_channel_fn, fake_quant_grad_per_channel_stub);
 
 } // namespace native
 } // namespace at