--- a/aten/src/ATen/native/quantized/fake_quant_affine.h
+++ b/aten/src/ATen/native/quantized/fake_quant_affine.h
@@ -5,7 +5,7 @@
 
 namespace at {
 
-class TensorIterator;
+struct TensorIterator;
 
 namespace native {