@supriyar 2020-02-26
This looks great! 
Since FP16 training is used on GPUs, the fake_quant nodes should support FP16 input/output types in the future. Would it be possible to generalize the type of the cuda kernels to work with FP16 in this PR?