@supriyar 2020-02-26
> @supriyar How about in next PR? It's better to do one single thing in a single PR. It should be very simple to generalize.

That sounds good. Thanks!